replaced dummy data with bd data
