--- src/app/test/page.tsx
@@ -1,12 +1,0 @@
-// src/app/test/page.tsx (just temporary)
-
-import { getTeams } from "@/data/teams";
-
-export default async function TestPage() {
-  const teams = await getTeams();
-  return (
-    <div>
-      <pre>{JSON.stringify(teams, null, 2)}</pre>
-    </div>
-  );
-}
