+++ src/constants/maps.ts
@@ -1,0 +1,4 @@
+export const LOCATIONS = {
+  HOME: { lat: 32.715736, lng: -117.161087 },
+  // Add other locations here
+} as const;
