+++ src/constants/ui.ts
@@ -1,0 +1,1 @@
+export const TABS = ["Popular", "Active Doors"];
